update to v2.0.0.0

+++ sus.py
@@ -8,20 +8,30 @@ async def suslogo(ctx):
 
 # susout command; a user accuses another user of something suspicious
 @bot.command()
-async def sus(ctx, arg1, *, arg2):
-    # deletes the command message
+async def sus(ctx, arg1='', *, arg2=''):
+    # checks for target user and reason and returns error if none
+    if not arg1:
+        await ctx.send("Please choose a target")
+        return
+    if not arg2:
+        await ctx.send("Please give a reason for the sus")
+        return
+    # checks if it is user (but this isn't the proper way to do it, so it will be fixed)
+    if not arg1.startswith("<"):
+        await ctx.send("Please choose a user to target")
+        return
+    # deletes the command message & sends sus message: format <user> sussed <user2> for <reason>
     await ctx.message.delete()
-    # gets the @ of the user that activated the command, adds the accused (arg1), followed by the action the accused did (arg2)
     await ctx.send('{0.author.mention} sussed {1} of *{2}*'.format(ctx, arg1, arg2))
 
 # susrate command; rates how sus a person is; is a static value
 @bot.command()
 async def susrate(ctx, *, msg=''):
-    # if there is is no message (ie its just s!susrate) it sets the user to be susrated as the user that activated the command
+    # if no follow up message after command, sets user to command user; otherwise user is who is in message
     if not msg:
         user = ctx.author.mention
-    # if a user is tagged, the user is susrated instead
     else:
+        # should probably make a way to error check this
         user = msg
     # generates a random number to use as a percent
     rate = random.randint(0, 100)
@@ -39,42 +49,30 @@ async def susimg(ctx, *, msg: discord.User=''):
         avatar = msg.display_avatar
     # sets it as a string so the code doesn't freak out
     pfp_url = str(avatar)
-    # checks if the profile picture is a gif
+    # checks if the profile picture is a gif and saves first frame of image; otherwise, save it as regular png
     if pfp_url[:-10].endswith("gif"):
-        # downloads image
         with requests.get(pfp_url) as r:
             img = r.content
-        # saves image
         with open('gifImage.gif', 'wb') as handler:
             handler.write(img)
-        # converts image into RGBA mode
         frame = Image.open("gifImage.gif").convert("RGBA")
-        # finds the first frame of gif
         frame.seek(0)
-        # saves the first frame as the image
         frame.save("image.png")
-    # if the profile picture is not a gif
     else:
-        # downloads image
         with requests.get(pfp_url) as r:
             img = r.content
-        # saves image
         with open('image.png', 'wb') as handler:
             handler.write(img)
-    # this is the image that "cuts" the profile picture
+    # opens among us character mask and uses it to cut user's avatar
     mask = Image.open("img/Red_body_mask.png").convert('L')
-    # this is the profile picture
     image = Image.open("image.png")
-    # this cuts the profile picture
     cut = ImageOps.fit(image, mask.size, centering=(0.5, 0.5))
     cut.putalpha(mask)
-    # this is the Among Us character iamge
+    # layers among us character image onto user's cut avatar
     susimg = Image.open('img/red_body.png')
-    # this combines the Among Us character with the cut profile picture
     susimg.paste(cut, (0, 0), cut)
-    # saves the image as final.png
+    # saves the image as final.png and sends image to server
     susimg.save('final.png')
-    # sends the image back
     await ctx.send(file=discord.File(open('final.png', 'rb'), 'final.png'))
     
 # different among us game feature commands below
@@ -82,8 +80,8 @@ async def susimg(ctx, *, msg: discord.User=''):
 # medbay command; medbay scans user
 @bot.command()
 async def scan(ctx):
-    user = ctx.author.mention
-    status = random.randint(0, 1)
+    # 20% chance of causing error
+    status = random.randint(0, 4)
     if status == 0:
         await ctx.send("Error: Cannot scan user")
     else:
